全部重构

+++ src/main/java/com/artland/util/PageResult.java
@@ -14,7 +14,7 @@ import java.util.List;
 public class PageResult implements Serializable {
 
 	//总记录数
-	private int totalCount;
+	private long totalCount;
 	//每页记录数
 	private int pageSize;
 	//总页数
@@ -32,7 +32,7 @@ public class PageResult implements Serializable {
 	 * @param pageSize   每页记录数
 	 * @param currPage   当前页数
 	 */
-	public PageResult(List<?> list, int totalCount, int pageSize, int currPage) {
+	public PageResult(List<?> list, long totalCount, int pageSize, int currPage) {
 		this.list = list;
 		this.totalCount = totalCount;
 		this.pageSize = pageSize;
@@ -40,7 +40,7 @@ public class PageResult implements Serializable {
 		this.totalPage = (int) Math.ceil((double) totalCount / pageSize);
 	}
 
-	public int getTotalCount() {
+	public long getTotalCount() {
 		return totalCount;
 	}
